r532 Добавиил в GuiMap метод clear. Сделал тест на 1  usercase Исправили ошибку с визой. В бд сейчас везде стоят 1. Не правильно. --------------------- r531 коммит чекбокса --------------------- r530
---------------------
r529
Лупа для карты!
---------------------
r528
GL-156 #comment Все сделано и привязано
---------------------
r527
bezier.js - подделаный алгоритм Эдика, для вывода кривулин полигоном на карту. Теперь и много, хоть и медленно. И не понятно что с zlevel. Кластеризация на сервере в этом коммите.
---------------------
r526
GL-190 #comment Доделал вывод туров и отелей
---------------------
r525
GL-190 Чего я только не сделал, сделал скролл, написал авиабилеты, прикрутил событие на колесико мышки и прочее прочее. Все выводиться, осталось сделать опрятный вид и убрать быдлятину в верстке и javascript
---------------------
r524
Иправил тест инаписол toString для некоторых классов
---------------------
r523
Сделал специальный пакет для интеграционных тестов
---------------------
r522
Исправил запрос и PropertySegment
---------------------
r520
GL-151 Исправил косяк с некрасивыми кривулинами, когда при маленьком спане они выгибались к бесконечности.
---------------------
r519
GL-151 Запили алгоритм.
---------------------
r518

---------------------
r517
Смержил. Исправил ошибки. Какая то хуета творится с тегом Alone.
---------------------
r516
Рабочий парсер, который создает лог файл в корневую папку сервера. Не понятно почему, но работает только из тестов функция dao.update(). Конкретно у меня лог парсера идёт в C:\Program Files (x86)\springsource\sts-2.8.1.RELEASE
---------------------
r513
GL-151 Почти доделал алгоритм. Осталось сделать нормальную функцию inPoligon.
---------------------
r512
Исправил косяк в запросе
---------------------
r511
Дбавил индексы для городов
---------------------
r510
дамб базы с игдексами
---------------------
r509
добавил дамп базы
---------------------
r508
Удалил new_page.
---------------------
r507
bezier.js - подделаный алгоритм Эдика, для вывода кривулин полигоном на карту
---------------------
r505
Исправил foodCost на cost
---------------------
r503
Исправил ошибки слияния
---------------------

+++ Globerry/src/main/java/com/globerry/project/service/gui/ISlider.java
@@ -13,13 +13,13 @@ import com.globerry.project.utils.PropertySegment;
  */
 public interface ISlider extends IGuiComponent{
     
-    float getLeftValue();
+    int getLeftValue();
     
-    float getRightValue();
+    int getRightValue();
     
-    void setLeftValue(float value);
+    void setLeftValue(int value);
     
-    void setRightValue(float value);     
+    void setRightValue(int value);     
     
     PropertyType getPropertyType();
